pyproject.toml
[project]
name = "PyBayes"
version = "0.0.1"
description = "Long lived dev repo for exploring different modelling packages"
requires-python = ">=3.12, <3.14"
dependencies = [
    "numpy",
    "pandas==2.3.2",
    "sympy",
    "scipy",
    "sqlalchemy",
    "matplotlib",
    "seaborn",
    "arviz",
    "tensorflow-macos>=2.16.2",
    "tensorflow-probability",
    "tensorflow-datasets",
    "torch",
    "torch-geometric",
    "transformers",
    "einops",
    "accelerate",
    "pymde",
    "scikit-learn",
    "statsmodels",
    "networkx",
    "jaxlib",
    "jax>=0.4.34",
    "jax-metal==0.1.1",
    "lifelines",
    "scikit-survival",
    #    "matlabengine==23.2.1",
    "flax",
    "blackjax",
    "optax",
    "cvxpy",
    "pzflow",
    "cvxpylayers",
    "antlr4-python3-runtime",
    "jupyter",
    "jupyterlab",
    "ipython",
    "ipywidgets",
    "jupytext",
    "kagglehub",
    "arxiv",
    "pymc",
    "simpy",
    "langchain",
    "keras==3.11.3",
    "streamlit",
    "psycopg2-binary",
    "mlx",
    "einops",
    "antlr4-tools",
    "open3d",
    "bambi",
    "pymc-marketing",
    "pymc-experimental",
    "sympytensor",
    "pysd",
    "datasets",
    "ydf==0.9.0",
    "pandas-stubs==2.3.0.250703",
    "kymatio==0.3.0",
    "boto3==1.40.20",
    "boto3-stubs==1.40.21",
    "python-dotenv==1.1.1",
    "dowhy>=0.12",
    "econml>=0.16.0",
    "guppylang>=0.20.0",
    "mcp>=1.16.0",
    "langchain-ollama>=0.3.10",
    "loguru>=0.7.3",
    "mcp-use>=1.3.11",
    "haystack-ai>=2.18.1",
    "ollama-haystack>=5.2.0",
    "mcp-haystack>=0.7.0",
    "chirho[extras]>=0.3.0",
    "numpyro>=0.19.0",
    "giotto-tda>=0.6.2",
]

[tool.uv]
package = false
